Consolidate diagnostic messages into one module
